+++ Dockerfile-builder
@@ -23,7 +23,7 @@ RUN python3 -m pip install --upgrade pip \
 ENV USER=builder
 ENV HOME=/home/${USER}
 ENV PATH=${HOME}/.local/bin:${PATH}
-ENV PYTHONPATH=${HOME}/rkvst_simplehash:${PYTHONPATH}
+ENV PYTHONPATH=${HOME}/datatrails_simplehash:${PYTHONPATH}
 RUN mkdir -p ${HOME}
 WORKDIR ${HOME}
