+++ scripts/builder.sh
@@ -30,5 +30,5 @@ docker run \
     -e TEST_CLIENT_SECRET_FILENAME \
     -e TEST_DEBUG \
     -e TEST_REFRESH_TOKEN \
-    rkvst-simplehash-python-builder \
+    datatrails-simplehash-python-builder \
     "$@"
