+++ scripts/functest.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 #
-# Manual testing of the rkvst_simplehashv1 command.
+# Manual testing of the datatrails_simplehashv1 command.
 #
 # 'task venv' to generate virtual environment
 #
@@ -17,33 +17,33 @@ API_QUERY+="&order_by=SIMPLEHASHV1"
 
 echo
 echo Help
-rkvst_simplehashv1 -h 
+datatrails_simplehashv1 -h 
 
 echo
 echo "Get using token file"
-if [ -n "${RKVST_AUTHTOKEN_FILENAME}" ]
+if [ -n "${DATATRAILS_AUTHTOKEN_FILENAME}" ]
 then
-    rkvst_simplehashv1 \
-        --auth-token-file "${RKVST_AUTHTOKEN_FILENAME}" \
+    datatrails_simplehashv1 \
+        --auth-token-file "${DATATRAILS_AUTHTOKEN_FILENAME}" \
         "${API_QUERY}"
 else
-    echo "Cannot use authtoken - please set RKVST_AUTHTOKEN_FILENAME"
+    echo "Cannot use authtoken - please set DATATRAILS_AUTHTOKEN_FILENAME"
 fi
 
 echo
 echo "Get using client id and secret"
-if [ -n "${RKVST_APPREG_CLIENT}" -a "${RKVST_APPREG_SECRET_FILENAME}" ]
+if [ -n "${DATATRAILS_APPREG_CLIENT}" -a "${DATATRAILS_APPREG_SECRET_FILENAME}" ]
 then
-    rkvst_simplehashv1 \
-        --client-id "${RKVST_APPREG_CLIENT}" \
-        --client-secret-file "${RKVST_APPREG_SECRET_FILENAME}" \
+    datatrails_simplehashv1 \
+        --client-id "${DATATRAILS_APPREG_CLIENT}" \
+        --client-secret-file "${DATATRAILS_APPREG_SECRET_FILENAME}" \
         "${API_QUERY}"
 else
-    echo "Cannot use appregistration - please set RKVST_APPREG_CLIENT"
-    echo "Cannot use appregistration - please set RKVST_APPREG_SECRET_FILENAME"
+    echo "Cannot use appregistration - please set DATATRAILS_APPREG_CLIENT"
+    echo "Cannot use appregistration - please set DATATRAILS_APPREG_SECRET_FILENAME"
 fi
 
-API_QUERY="https://app.dev-paul-0.wild.jitsuin.io"
+API_QUERY="https://app.datatrails.ai"
 API_QUERY+="/archivist/v2/publicassets/-/events"
 API_QUERY+="?proof_mechanism=SIMPLE_HASH"
 API_QUERY+="&timestamp_accepted_since=2022-10-07T07:01:34Z"
@@ -52,6 +52,6 @@ API_QUERY+="&order_by=SIMPLEHASHV1"
 
 echo
 echo "Get public assets"
-rkvst_simplehashv1 "${API_QUERY}"
+datatrails_simplehashv1 "${API_QUERY}"
 
 deactivate
