+++ scripts/unittests.sh
@@ -8,7 +8,7 @@ rm -f coverage.xml
 rm -rf htmlcov
 COVERAGE="coverage"
 ${COVERAGE} --version
-${COVERAGE} run --branch --source rkvst_simplehash -m unittest -v 
+${COVERAGE} run --branch --source datatrails_simplehash -m unittest -v 
 ${COVERAGE} html
 ${COVERAGE} xml
 ${COVERAGE} report
