+++ setup.cfg
@@ -4,15 +4,15 @@ statistics = True
 max-line-length = 88
 
 [metadata]
-name = rkvst-simplehash
-author = RKVST Inc.
-author_email = support@rkvst.com
-description = RKVST Simplehash algorithm for DLT Anchoring
+name = datatrails-simplehash
+author = DataTrails Inc.
+author_email = support@datatrails.ai
+description = DataTrails Simplehash algorithm for DLT Anchoring
 license = MIT
 license_files = LICENSE
 long_description = file: README.rst
 long_description_content_type = text/x-rst
-url = https://github.com/rkvst/rkvst-simplehash-python
+url = https://github.com/datatrails/datatrails-simplehash-python
 
 classifiers =
     Development Status :: 4 - Beta
@@ -24,20 +24,20 @@ classifiers =
     Topic :: Utilities
 
 project_urls =
-    Source = https://github.com/rkvst/rkvst-simplehash-python
-    Tracker = https://github.com/rkvst/rkvst-simplehash-python/issues
+    Source = https://github.com/datatrails/datatrails-simplehash-python
+    Tracker = https://github.com/datatrails/datatrails-simplehash-python/issues
 
 [options]
 install_requires = file: requirements.txt
-packages = rkvst_simplehash
+packages = datatrails_simplehash
 platforms = any
 python_requires = >=3.8
 setup_requires = setuptools-git-versioning
 
 [options.entry_points]
 console_scripts =
-    rkvst_simplehashv1 = rkvst_simplehash.v1:main
-    rkvst_simplehashv2 = rkvst_simplehash.v2:main
+    datatrails_simplehashv1 = datatrails_simplehash.v1:main
+    datatrails_simplehashv2 = datatrails_simplehash.v2:main
 
 [bdist_wheel]
 universal = 1
